requirements.txt
streamlit>=1.28.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0